feat: 엔티티 수정할부분 + 일부 디티오 객체 생성

+++ src/main/java/com/dain_review/domain/post/model/entity/PostMeta.java
@@ -21,6 +21,6 @@ public class PostMeta extends BaseEntity {
     @JoinColumn(name = "post_id")
     private Post post;
 
-    private Integer viewCount;
-    private Integer commentCount;
+    private Long viewCount;
+    private Long commentCount;
 }
